Change change of To_String
The RFC 4122 standard specifies that the hexidecimal representation of
the UUID is in lowercase characters, not uppercase.

"Each field is treated as an integer and has its value printed as a
zero-filled hexadecimal digit string with the most significant
digit first.  The hexadecimal values "a" through "f" are output as
lower case characters and are case insensitive on input."